+++ detectors/aws/ecs/test/metadatav4-response-container-ec2.json
