+++ detectors/aws/ecs/test/metadatav4-response-container-fargate.json
