+++ detectors/aws/ecs/test/metadatav4-response-task-ec2.json
